+++ src/frykit/plot/api.py
@@ -35,7 +35,7 @@ from shapely.geometry.base import BaseGeometry
 
 from frykit import get_data_dirpath
 from frykit.calc import asarrays, get_values_between, lon_to_180
-from frykit.configuration import DataSource, config
+from frykit.conf import DataSource, config
 from frykit.plot.artist import (
     Compass,
     Frame,